move gameservice to lifehack library

+++ Project/StaticLibrary/LifeHack/IOCP.h
@@ -6,7 +6,7 @@ public:
 	IOCP();
 	~IOCP();
 
-	void Init( int Port );
+	void Init( short Port );
 	void Release();
 
 private:
